+++ plone/formwidget/geolocation/widget.py
@@ -11,13 +11,11 @@ from z3c.form.interfaces import IFieldWidget
 from z3c.form.interfaces import IFormLayer
 from z3c.form.widget import FieldWidget
 from zope.component import adapter
-from zope.component import queryMultiAdapter
 from zope.i18n import translate
 from zope.interface import implementer
 from zope.interface import implementer_only
 
 import json
-import six
 
 
 @implementer_only(IGeolocationWidget)
@@ -52,7 +50,7 @@ class GeolocationWidget(TextWidget):
         """Return the geo location as GeoJSON string.
         """
         coordinates = self.value
-        if not coordinates:
+        if not all(coordinates):
             return
 
         title = getattr(self.context, "title", "") or ""
@@ -98,10 +96,11 @@ class GeolocationWidget(TextWidget):
             "geosearch_provider": getrec("geolocation.geosearch_provider"),
             "default_map_layer": getrec("geolocation.default_map_layer"),
             "map_layers": [
-                {"title": translate(_(l), context=self.request), "id": l}
-                for l in map_layers
+                {"title": translate(_(layer), context=self.request), "id": layer}
+                for layer in map_layers
             ],
-            "default_location": self._default_loc(),
+            "latitude": getrec("geolocation.default_latitude") or None,
+            "longitude": getrec("geolocation.default_longitude") or None,
         }
         if self.mode == "input":
             # geosearch for input is always active
@@ -112,8 +111,8 @@ class GeolocationWidget(TextWidget):
 
     def _default_loc(self):
         return (
-            getrec("geolocation.default_latitude", default=0.0),
-            getrec("geolocation.default_longitude", default=0.0),
+            getrec("geolocation.default_latitude") or None,
+            getrec("geolocation.default_longitude") or None,
         )