+++ CHANGES.rst
@@ -1,7 +1,7 @@
 Changelog
 =========
 
-2.2.3 (unreleased)
+2.2.3 (2021-05-26)
 ------------------
 
 - Add plone.restapi serializer (if plone.restapi is installed).